handle error in driver

+++ lib/futures/setting/controller/setting_controller.dart
@@ -1,10 +1,8 @@
 import 'dart:convert';
 import 'dart:io';
-import 'package:dartz/dartz_unsafe.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:http/http.dart' as http;
-import 'package:image_picker/image_picker.dart';
 import 'package:selivery_controlle_panal/core/contants/api.dart';
 import 'package:selivery_controlle_panal/core/functions/internet_checker.dart';
 import 'package:selivery_controlle_panal/core/widgets/image_picker.dart';
@@ -23,6 +21,7 @@ class CategoryController extends GetxController {
   Future<void> getAllCategories() async {
     try {
       categoryLoading.value = true;
+      categoryList.value = <CategoryModel>[];
       final response =
           await http.get(allCategory, headers: authHeadersWithToken(token));
       if (response.statusCode == 200) {
@@ -30,7 +29,9 @@ class CategoryController extends GetxController {
         result['categories'].forEach((json) {
           categoryList.add(CategoryModel.fromJson(json));
         });
-        categoryError.value = '';
+        categoryList.isEmpty
+            ? categoryError.value = 'لا يوجد بيانات'
+            : categoryError.value = '';
         categoryLoading.value = false;
       } else {
         final result = json.decode(response.body);
@@ -90,6 +91,8 @@ class CategoryController extends GetxController {
         if (response.statusCode == 200 || response.statusCode == 201) {
           showDialogWithGetX("تم الاضافه بنجاح");
           addLoading.value = false;
+          clearData();
+          getAllCategories();
         } else {
           showDialogWithGetX(responsebody['message']);
           addLoading.value = false;
@@ -152,19 +155,30 @@ class CategoryController extends GetxController {
   }
 
   void deleteCategory(String? id) async {
-    try {
-      print(deleteCategoryUri(id));
-      final response = await http.delete(deleteCategoryUri(id),
-          headers: authHeadersWithToken(token));
-      final result = jsonDecode(response.body);
-      if (response.statusCode == 200) {
-        update();
-      } else {
-        print(result['message']);
+    if (await checkInternet()) {
+      try {
+        print(deleteCategoryUri(id));
+        final response = await http.delete(deleteCategoryUri(id),
+            headers: authHeadersWithToken(token));
+        final result = jsonDecode(response.body);
+        if (response.statusCode == 200) {
+          showDialogWithGetX(result['message']);
+          getAllCategories();
+        } else {
+          showDialogWithGetX(result['message']);
+        }
+      } catch (error) {
+        showDialogWithGetX(error.toString());
       }
-    } catch (error) {
-      print(error.toString());
+    } else {
+      showDialogWithGetX("لا يوجد اتصال بالانترنت");
     }
+  }
+
+  clearData() {
+    titleController.clear();
+    categoryImage = null;
+    update();
   }
 
   @override
@@ -177,6 +191,7 @@ class CategoryController extends GetxController {
   @override
   void dispose() {
     titleController.dispose();
+    clearData();
     // TODO: implement dispose
     super.dispose();
   }
